Add player join room broadcast

+++ Lockstep/Client/Player.cs
@@ -5,7 +5,7 @@ namespace Lockstep.Client;
 internal class Player
 {
     public required PlayerInfo Info { get; set; }
-    public int PortNumber { get; init; }
+    public byte PortNumber { get; init; }
     public DateTime LastSeen { get; set; } = DateTime.UtcNow;
     public Dictionary<ushort, PendingPacket> PendingPackets = [];
 }
